delete role feature

+++ jbaccess/web_ui/static/js/main.js
@@ -1,30 +1,30 @@
+let removal = [
+    {name: 'person', path: '/person'},
+    {name: 'key', path: '/keys'},
+    {name: 'role', path: '/roles'}
+];
+
 $(document).ready( () => {
     $('.clear').click( (event) => {
         $(event.target).parents('form').find('input').val('');
         Materialize.updateTextFields();
     });
 
-    $('.delete-person').click( (event) => {
-        let id = $(event.currentTarget).attr('data-id');
-
-        $.ajax({
-            url: `/person/${id}`,
-            type: 'DELETE',
-            success: result => {
-               window.location.replace(window.location.href)
-            }
-        });
-    });
-
-    $('.delete-key').click( (event) => {
-        let id = $(event.currentTarget).attr('data-id');
-
-        $.ajax({
-            url: `/keys/${id}`,
-            type: 'DELETE',
-            success: result => {
-               window.location.replace(window.location.href)
-            }
-        });
-    })
+    removal.forEach(element => deleteItem(element));
 });
+
+function deleteItem(item) {
+    let $element = $('.delete-' + item.name);
+    if ($element)
+        $element.click((event) => {
+            let id = $(event.currentTarget).attr('data-id');
+
+            $.ajax({
+                url: `${item.path}/${id}`,
+                type: 'DELETE',
+                success: result => {
+                    window.location.replace(window.location.href)
+                }
+            });
+        })
+}
